+++ src/AppBundle/DataFixtures/ORM/LoadExampleComments.php
@@ -6,6 +6,7 @@ use Doctrine\Common\DataFixtures\FixtureInterface;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Raddit\AppBundle\Entity\Comment;
+use Raddit\AppBundle\Entity\Submission;
 use Raddit\AppBundle\Entity\Url;
 use Raddit\AppBundle\Entity\User;
 use Raddit\AppBundle\Utils\MarkdownConverter;
@@ -15,7 +16,7 @@ class LoadExampleComments implements FixtureInterface, OrderedFixtureInterface {
      * {@inheritdoc}
      */
     public function load(ObjectManager $manager) {
-        $submission = $manager->getRepository(Url::class)->findOneBy([]);
+        $submission = $manager->getRepository(Submission::class)->findOneBy([]);
         $user = $manager->getRepository(User::class)->findOneByUsername('emma');
 
         $comment = Comment::create($submission, $user);