+++ src/AppBundle/DataFixtures/ORM/LoadExampleSubmissions.php
@@ -6,8 +6,7 @@ use Doctrine\Common\DataFixtures\FixtureInterface;
 use Doctrine\Common\DataFixtures\OrderedFixtureInterface;
 use Doctrine\Common\Persistence\ObjectManager;
 use Raddit\AppBundle\Entity\Forum;
-use Raddit\AppBundle\Entity\Post;
-use Raddit\AppBundle\Entity\Url;
+use Raddit\AppBundle\Entity\Submission;
 use Raddit\AppBundle\Entity\User;
 use Raddit\AppBundle\Utils\MarkdownConverter;
 
@@ -19,16 +18,23 @@ class LoadExampleSubmissions implements FixtureInterface, OrderedFixtureInterfac
         $forum = $manager->getRepository(Forum::class)->findOneBy(['name' => 'liberalwithdulledge']);
         $user = $manager->getRepository(User::class)->findOneByUsername('emma');
 
-        $url = Url::create($forum, $user);
+        $url = Submission::create($forum, $user);
         $url->setTitle('This is a submitted URL');
         $url->setUrl('http://www.example.com');
         $manager->persist($url);
 
-        $post = Post::create($forum, $user);
+        $post = Submission::create($forum, $user);
         $post->setTitle('This is a test submission');
         $post->setRawBody('Hi');
         $post->setBody(MarkdownConverter::convert($post->getRawBody()));
         $manager->persist($post);
+
+        $combo = Submission::create($forum, $user);
+        $combo->setTitle('This is a submission with both a URL and a body');
+        $combo->setUrl('http://www.example.org/some_thing_here');
+        $combo->setRawBody('The quick brown fox jumped over the lazy dog.');
+        $combo->setBody(MarkdownConverter::convert($combo->getRawBody()));
+        $manager->persist($combo);
 
         $manager->flush();
     }